// src/app/model/employee.ts
export interface Employee {


    empId : number;
    empName: string;         
    position: string;        
    empDepartment: string;  
    empsalary: number;     
    joinDate: string;         
    userId: number;           
    deductionId: number;      
    benefitId: number;        
    roleId: number;           
    managerId?: number;   
    
    
  }
  